Fix issues caused by merge conflicts
