+++ src/util/CancelablePromise.tsx
@@ -9,23 +9,6 @@
  * Please see LICENSE file for your rights under this license. */
 
 /**
- * Get the base path of the web interface. The API will inject a base element
- * for this purpose, but if the web interface is not hosted by the API, it will
- * fall back to the public URL set by Create React App.
- *
- * @returns The base path to use
- */
-export const getBasePath = (): string => {
-  const baseElement = document.getElementsByTagName("base")[0];
-
-  if (baseElement) {
-    return new URL(baseElement.href).pathname;
-  } else {
-    return process.env.PUBLIC_URL;
-  }
-};
-
-/**
  * A promise which can be canceled
  */
 export interface CancelablePromise<T> {
